Secure-Exam-Tool/secure_exam_tool/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
Pillow==10.0.1
Werkzeug==2.3.7 